+++ example/lib/main.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
 
-import 'package:flutter/services.dart';
+import 'package:selector/selector.dart';
 import 'package:functional_google_mobile_ads/functional_google_mobile_ads.dart';
-import 'package:google_mobile_ads/google_mobile_ads.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -19,35 +17,17 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
-  final _functionalGoogleMobileAdsPlugin = FunctionalGoogleMobileAds();
+  final _rewarded = FunctionalRewardedAd();
 
   @override
   void initState() {
     super.initState();
-    initPlatformState();
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    // We also handle the message potentially returning null.
-    try {
-      platformVersion =
-          await _functionalGoogleMobileAdsPlugin.getPlatformVersion() ?? 'Unknown platform version';
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
+  @override
+  void dispose() {
+    super.dispose();
+    _rewarded.dispose();
   }
 
   @override
@@ -55,12 +35,80 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('Functional Google Mobile Ads example'),
         ),
-        body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            ElevatedButton(
+                onPressed: () async {
+                  await FunctionalInterstitialAd.loadAndShow(
+                      adUnitId: TestAdId.interstitial);
+
+                  print('interstitial ok');
+                },
+                child: const Text('InterstitialAd')),
+            ElevatedButton(
+                onPressed: () async {
+                  final item = await FunctionalRewardedAd.loadAndShow(
+                      adUnitId: TestAdId.rewarded);
+
+                  print('item is ${item.amount}');
+                  print('item is ${item.type}');
+                },
+                child: const Text('RewardedAd')),
+            ElevatedButton(
+                onPressed: () async {
+                  final item = await FunctionalRewardedInterstitialAd.loadAndShow(
+                      adUnitId: TestAdId.rewardedInterstitial);
+
+                  print('item is ${item.amount}');
+                  print('item is ${item.type}');
+                },
+                child: const Text('RewardedInterstitialAd')),
+            const Spacer(),
+            FunctionalBannerAd(
+              bannerAdUnitId: TestAdId.banner,
+              adSize: AdSize.banner,
+            ),
+          ],
         ),
       ),
     );
   }
+}
+
+class TestAdId {
+  static final String appOpening = androidOrElse(
+    'ca-app-pub-3940256099942544/3419835294',
+    'ca-app-pub-3940256099942544/5662855259',
+  );
+  static final String banner = androidOrElse(
+    'ca-app-pub-3940256099942544/6300978111',
+    'ca-app-pub-3940256099942544/2934735716',
+  );
+  static final String interstitial = androidOrElse(
+    'ca-app-pub-3940256099942544/1033173712',
+    'ca-app-pub-3940256099942544/4411468910',
+  );
+  static final String interstitialVideo = androidOrElse(
+    'ca-app-pub-3940256099942544/8691691433',
+    'ca-app-pub-3940256099942544/5135589807',
+  );
+  static final String rewarded = androidOrElse(
+    'ca-app-pub-3940256099942544/5224354917',
+    'ca-app-pub-3940256099942544/1712485313',
+  );
+  static final String rewardedInterstitial = androidOrElse(
+    'ca-app-pub-3940256099942544/5354046379',
+    'ca-app-pub-3940256099942544/6978759866',
+  );
+  static final String nativeAdvanced = androidOrElse(
+    'ca-app-pub-3940256099942544/2247696110',
+    'ca-app-pub-3940256099942544/3986624511',
+  );
+  static final String nativeAdvancedVideo = androidOrElse(
+    'ca-app-pub-3940256099942544/1044960115',
+    'ca-app-pub-3940256099942544/2521693316',
+  );
 }