+++ lib/functional_google_mobile_ads.dart
@@ -1,8 +1,5 @@
-
-import 'functional_google_mobile_ads_platform_interface.dart';
-
-class FunctionalGoogleMobileAds {
-  Future<String?> getPlatformVersion() {
-    return FunctionalGoogleMobileAdsPlatform.instance.getPlatformVersion();
-  }
-}
+export 'package:google_mobile_ads/google_mobile_ads.dart';
+export 'package:functional_google_mobile_ads/functional_banner_ad.dart';
+export 'package:functional_google_mobile_ads/functional_rewarded_ad.dart';
+export 'package:functional_google_mobile_ads/functional_rewarded_interstitial_ad.dart';
+export 'package:functional_google_mobile_ads/functional_interstitial_ad.dart';