+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 0.0.1
+## 0.1.5
+- unspecify `google_mobile_ads` and `plugin_platform_interface` version
 
-* TODO: Describe initial release.
+## 0.1.4
+* dispose finally
+
+## 0.1.3
+* FunctionalInterstitialAd
+* FunctionalRewardedAd
+* FunctionalRewardedInterstitialAd
+* FunctionalBannerAd
+
+## 0.0.2
+* FunctionalAdmobRewarded
+
+## 0.0.1
+* FunctionalAdmobBanner